requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
cloudinary==1.15.0
dj-database-url==0.5.0
Django==2.1.4
django-crispy-forms==1.7.2
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.8
meetup-api==0.1.1
mock==2.0.0
pbr==5.1.2
psycopg2==2.7.7
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
whitenoise==4.1.2
